requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
gradio==5.23.2
gradio_client==1.8.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.11
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
orjson==3.11.2
packaging==23.2
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.31.0
requests-toolbelt==1.0.0
rich==14.1.0
ruff==0.12.9
safehttpx==0.1.6
safetensors==0.6.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.15.2
tomlkit==0.13.3
torch==2.2.1
tqdm==4.67.1
transformers==4.38.2
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
